+++ src/interface/IOracle.sol
@@ -3,6 +3,6 @@ pragma solidity ^0.8.0;
 
 interface IOracle {
     function isOutcomeSet() external view returns (bool);
-    function getOutcome() external view returns (int);
-    function setOutcome (int _outcome) external;
+    function getOutcome() external view returns (int256);
+    function setOutcome(int256 _outcome) external;
 }
